optimize the code to support expandable item without adding new widget
