+++ lib/src/inherited_nested_choice_list_view.dart
@@ -4,13 +4,13 @@ import 'package:nested_choice_list/src/nested_choice_entity.dart';
 /// An inherited widget that provides access to the selected items in a
 /// nested list view.
 ///
-/// The [InheritedNestedNavigationListView] widget allows descendant widgets to
+/// The [InheritedNestedChoiceListView] widget allows descendant widgets to
 /// access the set of selected items and listen for changes.
-class InheritedNestedNavigationListView extends InheritedWidget {
-  /// Creates an [InheritedNestedNavigationListView] widget.
+class InheritedNestedChoiceListView extends InheritedWidget {
+  /// Creates an [InheritedNestedChoiceListView] widget.
   ///
   /// The [selectedItems] parameter must not be null.
-  const InheritedNestedNavigationListView({
+  const InheritedNestedChoiceListView({
     required super.child,
     required this.selectedItems,
     super.key,
@@ -19,19 +19,19 @@ class InheritedNestedNavigationListView extends InheritedWidget {
   /// The set of selected items.
   final Set<NestedChoiceEntity> selectedItems;
 
-  /// Retrieves the nearest [InheritedNestedNavigationListView] widget up the
+  /// Retrieves the nearest [InheritedNestedChoiceListView] widget up the
   /// widget tree.
   ///
   /// This method allows descendant widgets to access the
-  /// [InheritedNestedNavigationListView] and its selected items.
-  static InheritedNestedNavigationListView? of(BuildContext context) {
-    return context.dependOnInheritedWidgetOfExactType<
-        InheritedNestedNavigationListView>();
+  /// [InheritedNestedChoiceListView] and its selected items.
+  static InheritedNestedChoiceListView? of(BuildContext context) {
+    return context
+        .dependOnInheritedWidgetOfExactType<InheritedNestedChoiceListView>();
   }
 
   @override
   bool updateShouldNotify(
-    covariant InheritedNestedNavigationListView oldWidget,
+    covariant InheritedNestedChoiceListView oldWidget,
   ) {
     return selectedItems.length != oldWidget.selectedItems.length;
   }
